+++ lib/inappstory_sdk_module.dart
@@ -2,6 +2,12 @@ import 'dart:async';
 
 /// This is an abstract class that defines the interface for initializing the InAppStory SDK module.
 abstract class InAppStorySdkModule {
-  FutureOr<void> initWith(String apiKey, String userId,
-      {String? languageCode, String? languageRegion});
+  FutureOr<void> initWith(
+    String apiKey,
+    String userId, {
+    String? languageCode,
+    String? languageRegion,
+    String? userSign,
+    String? cacheSize,
+  });
 }