+++ pigeons/pigeon_generation.dart
@@ -18,8 +18,14 @@ import 'package:pigeon/pigeon.dart';
 @HostApi()
 abstract class InappstorySdkModuleHostApi {
   @async
-  void initWith(String apiKey, String userID,
-      {String? languageCode, String? languageRegion});
+  void initWith(
+    String apiKey,
+    String userID, {
+    String? userSign,
+    String? languageCode,
+    String? languageRegion,
+    String? cacheSize,
+  });
 
   void createListAdaptor(String feed);
 
@@ -33,7 +39,9 @@ abstract class InAppStoryManagerHostApi {
   void setTags(List<String> tags);
 
   @async
-  void changeUser(String userId);
+  void changeUser(String userId, {String? userSign});
+
+  void userLogout();
 
   void closeReaders();